% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/L0glm.R
\name{control.l0.gen}
\alias{control.l0.gen}
\alias{control.iwls.gen}
\alias{control.fit.gen}
\title{Set parameters for L0glm algorithm}
\usage{
control.l0.gen(maxit = 100, rel.tol = 1e-04, delta = 1e-05,
  gamma = 2, warn = FALSE)

control.iwls.gen(maxit = 100, rel.tol = 1e-04, thresh = 1e-05,
  warn = FALSE)

control.fit.gen(maxit = 10, block.size = NULL, tol = 1e-07)
}
\arguments{
\item{maxit}{The maximum number of iterations.}

\item{rel.tol}{Coefficients are update until the relative difference between two iteration
is smaller than this threshold. The relative difference is defined as
\eqn{2 \frac{|coef_i^{(k-1)} - coef_i^{(k)}|}{coef_i^{(k-1)} + coef_i^{(k)}}}}

\item{delta}{A small constant added to the estimated coefficients when computing the
adaptive ridge penalty weights.}

\item{gamma}{The exponent used for computing the adaptive ridge penalty weights. It is
recommended to work with \code{gamma = 2} (Frommlet et Nuel 2016).}

\item{warn}{Should warnings from the iteration be produced ?}

\item{thresh}{A threshold under which the fitted coefficients are clipped to 0.}

\item{block.size}{Number of covariates to fit simultaneously. If \code{block.size} is smaller
than \code{p}, the fit is performed in different blocks which is updated
cyclically.}

\item{tol}{Stop iteration when the difference in coefficients between iterations is
smaller than this threshold. The coefficient difference is defined as:
\eqn{\frac{\|coefs^{(k-1)} - coefs^{(k)}}{\|coefs^{(k-1)}\|}}}
}
\value{
The functions return a list with the required parameters for computing an
\code{L0glm} fit.
}
\description{
The functions generate a list with all required parameters for the control
the adaptive ridge iteration (\code{control.l0.gen}), the iteratively
weighted least squares for GLM fitting (\code{control.iwls.gen}), or the
least squares fitting (\code{control.fit.gen}).
}
\section{Functions}{
\itemize{
\item \code{control.iwls.gen}: Generate parameters that control the IWLS iteration.

\item \code{control.fit.gen}: Generates parameters that control the least square fit.
}}

\examples{

control.l0.gen()
control.iwls.gen()
control.fit.gen()

}
\references{
Frommlet, F. and G., Nuel. (2016) An adaptive ridge procedure for L0 regularization. PloS one.
}
\seealso{
\code{\link{L0glm}}
}
