\name{l0glm-package}
\alias{l0glm-package}
\alias{l0glm}
\docType{package}
\title{
\packageTitle{l0glm}
}
\description{
\packageDescription{l0glm}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{l0glm}
\packageIndices{l0glm}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{l0glm}

Maintainer: \packageMaintainer{l0glm}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
