% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_benchmark}
\alias{plot_benchmark}
\title{Plot benchmark examples}
\usage{
plot_benchmark(x, y, fit, inference = NULL, beta_true, legend = FALSE,
  ...)
}
\arguments{
\item{x}{a vector of length \code{n} with time labels.}

\item{y}{a vector of length \code{n} with signal.}

\item{fit}{an \code{L0glm} object}

\item{inference}{the output of \code{L0glm.inference}}

\item{...}{further arguments passed to \code{plot}}

\item{a.true}{the true coefficients}
}
\value{
The function returns \code{NULL}, but plots \code{y} with the expected
coefficients and the fitted values from \code{fit} with the estimated
coefficients in a way that gives an intuitive view on the quality of the fit.
}
\description{
Function to plot the comparison between the fitted model and the known true
model. It used in the \code{\link{L0glm}} examples and should not be expected
to useful in another use.
}
